pyproject.toml
[project]
name = "scribeauth"
version = "1.2.0"
description = "Library to authenticate to Scribe's platform"
readme = "README.md"
license = "LICENSE.md"
requires-python = ">=3.11"
dependencies = [
    "boto3[cognito-idp] (>=1.37.11,<2.0.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)"
]
keywords = [
    'Development Status :: 4 - Beta',
    'Intended Audience :: Developers',
    'Operating System :: OS Independent',
    'License :: OSI Approved :: MIT License',
    'Natural Language :: English',
    'Programming Language :: Python :: 3.11',
    'Topic :: Security',
    'Typing :: Typed'
]
url = 'https://github.com/ScribeLabsAI/ScribeAuth'


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
types-boto3 = {extras = ["cognito-idp"], version = "^1.37.11"}
pytest = "^8.3.5"
sphinx = "^8.2.3"
pyotp = "^2.9.0"
dotenv = "^0.9.9"
sphinx-markdown-builder = "^0.6.8"
coverage = "^7.6.12"

